# memhotplug.py: optimised to run for user specified memory ratio