Test was unplugging whole free memroy making the system go hung, so
now user has option to specify % of memory to unplug. by default we have
5 % memory. Now we do not have OOM killers being called while test.

Signed-off-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>

JOB ID     : 86da0d47189f331471727f3a1128d2f44947639b
JOB LOG    : /root/avocado-korg/korg/results/job-2018-04-12T06.05-86da0d4/job.log
 (1/4) memhotplug.py:memstress.test_hotplug_loop: PASS (205.33 s)
 (2/4) memhotplug.py:memstress.test_hotplug_toggle: PASS (104.39 s)
 (3/4) memhotplug.py:memstress.test_dlpar_mem_hotplug: PASS (116.04 s)
 (4/4) memhotplug.py:memstress.test_hotplug_per_numa_node: PASS (125.34 s)
RESULTS    : PASS 4 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 570.22 s
JOB HTML   : /root/avocado-korg/korg/results/job-2018-04-12T06.05-86da0d4/html/results.html

[memhotpluglogs.txt](https://github.com/avocado-framework-tests/avocado-misc-tests/files/1902720/memhotpluglogs.txt)